fixed a ton of ... bugs

+++ implementation/gps-sleeper/src/pt/isec/gps/g22/sleeper/ui/SetAlarmActivity.java
@@ -53,9 +53,9 @@ public class SetAlarmActivity extends Activity {
         cal.set(Calendar.MINUTE,min);
         cal.set(Calendar.SECOND, 0);    
 		
-        int sleepDate,wakeDate;
-        sleepDate = (int) millis / 1000;
-        wakeDate = (int) cal.getTimeInMillis() / 1000;
+        long sleepDate,wakeDate;
+        sleepDate = millis / 1000;
+        wakeDate = cal.getTimeInMillis() / 1000;
         //Insert daily record
 		dayRecord = new DayRecord(sleepDate,wakeDate);
 		dayRecord.setExhaustion((int)rbExhaustion.getRating());
